requirements.txt
discord.py==2.1.0
openai==0.19.0
python-dotenv==0.21.0
